[+] : as a verb
dyadic : will set x to y if x is a variable.
otherwise it will return y

monadic : is now the identity function. As a result
monadic + is a placeholder for 'flip'

minor change in the represent adverbs in the parse tree.
adverbs are now represented as their index to the 'va'
function pointer LUT+an offset which allows us to
disambiguate between adverbs and verbs